Revert "Ui changes 3-0-stable"
