+++ Gemfile
@@ -1,3 +1,3 @@
 source 'https://rubygems.org'
-gem 'spree', github: 'spree/spree', branch: '3-0-stable'
+gem 'spree', github: 'spree/spree', branch: '2-2-stable'
 gemspec
