+++ app/controllers/spree/checkout_controller_decorator.rb
@@ -2,23 +2,22 @@ Spree::CheckoutController.class_eval do
 
   Spree::PermittedAttributes.checkout_attributes << :gift_code
 
-  def update
-    if @order.update_from_params(params, permitted_checkout_attributes, request.headers.env)
+  durably_decorate :update, mode: 'soft', sha: '908d25b70eff597d32ddad7876cd89d3683d6734' do
+    if @order.update_from_params(params, permitted_checkout_attributes)
       if @order.gift_code.present?
         render :edit and return unless apply_gift_code
       end
 
-      @order.temporary_address = (params[:save_user_address]).blank?
-
+      persist_user_address
       unless @order.next
         flash[:error] = @order.errors.full_messages.join("\n")
-        redirect_to checkout_state_path(@order.state)
+        redirect_to checkout_state_path(@order.state) and return
       end
 
       if @order.completed?
-        @current_order = nil
+        session[:order_id] = nil
         flash.notice = Spree.t(:order_processed_successfully)
-        flash['order_completed'] = true
+        flash[:commerce_tracking] = "nothing special"
         redirect_to completion_route
       else
         redirect_to checkout_state_path(@order.state)
