+++ app/models/spree/gift_card.rb
@@ -2,7 +2,6 @@ require 'spree/core/validators/email'
 
 module Spree
   class GiftCard < ActiveRecord::Base
-    include CalculatedAdjustments
 
     UNACTIVATABLE_ORDER_STATES = ["complete", "awaiting_return", "returned"]
 
@@ -18,7 +17,10 @@ module Spree
     validates :original_value,     presence: true
 
     before_validation :generate_code, on: :create
+    before_validation :set_calculator, on: :create
     before_validation :set_values, on: :create
+
+    include Spree::Core::CalculatedAdjustments
 
     def apply(order)
       # Nothing to do if the gift card is already associated with the order
@@ -61,10 +63,6 @@ module Spree
       !UNACTIVATABLE_ORDER_STATES.include?(order.state)
     end
 
-    def calculator
-      Spree::Calculator::GiftCardCalculator.new
-    end
-
     private
 
     def generate_code
@@ -73,6 +71,9 @@ module Spree
       end
     end
 
+    def set_calculator
+      self.calculator = Spree::Calculator::GiftCard.new
+    end
 
     def set_values
       self.current_value  = self.variant.try(:price)
