+++ app/models/spree/payment_decorator.rb
@@ -1,13 +1,9 @@
 Spree::Payment.class_eval do
-  after_update :send_gift_card, if: :gift_card_valid_state_transition?
+  before_update :send_gift_card, :if => "state_changed? && state_was != 'completed' && state == 'completed'"
 
-  def send_gift_card 
+  def send_gift_card
     order.line_items.each do |li|
       Spree::OrderMailer.gift_card_email(li.gift_card.id, order).deliver if li.gift_card
     end
-  end
-
-  def gift_card_valid_state_transition?
-    state_changed? && state_was != 'completed' && state == 'completed'
   end
 end
