+++ app/overrides/spree/checkout/_payment/_gift_code_field.html.erb.deface
@@ -1,7 +1,7 @@
-<!-- insert_bottom '[data-hook=checkout_payment_step]'
+<!-- insert_bottom '#payment'
      original '196c020541e91ab34bec98a975e606f0305ebb2b' -->
 
 <p class='field' data-hook='gift_code'>
   <%= form.label :gift_code %><br />
-  <%= form.text_field :gift_code, value: nil, class: 'form-control' %>
+  <%= form.text_field :gift_code, value: nil %>
 </p>
