+++ app/overrides/spree/layouts/admin/gift_cards_tab.html.erb.deface
@@ -2,8 +2,4 @@
      enabled
      original 'fc9c9d3ac8a3ce2ae5deb141d683fde887e4ff75' -->
 
-<ul class="nav nav-sidebar">
-  <li class="sidebar-menu-item">
-    <%= tab(:gift_cards, spree.admin_gift_cards_path, icon: 'gift') %>
-  </li>
-</ul>
+<%= tab(:gift_cards, :url => spree.admin_gift_cards_path) %>
