+++ app/overrides/spree/orders/_form/add_gift_code_field.html.erb.deface
@@ -3,11 +3,9 @@
 <tbody class="gift-code-field">
   <tr>
     <td colspan="6">
-    <div class='form-inline'>
       <%= order_form.label :gift_code %>
-      <%= order_form.text_field :gift_code, size: 30, class: 'form-control' %>
-      <%= order_form.submit Spree.t(:apply), class: 'btn btn-primary' %>
-      </div>
+      <%= order_form.text_field :gift_code, :size => 10 %>
+      <%= order_form.submit Spree.t(:apply) %>
     </td>
   </tr>
 </tbody>
