+++ app/overrides/spree/shared/_sidebar/insert_bottom_sidebar.html.erb.deface
@@ -2,7 +2,4 @@
      enabled
      original '2f11ce271ae3b346b9fc6a927598ad6d6d6a1885' -->
 
-<div class="list-group">
-  <%= link_to Spree.t("buy_gift_card"), new_gift_card_path, class: 'list-group-item' %>
-</div>
-
+<%= link_to Spree.t("buy_gift_card"), new_gift_card_path, :class => 'button' %>
