+++ app/views/spree/admin/gift_cards/index.html.erb
@@ -3,14 +3,16 @@
 <% end %>
 
 <% content_for :page_actions do %>
-  <%= button_link_to Spree.t(:new_gift_card), new_object_url, :icon => 'plus', class: 'btn btn-success' %>
+  <li>
+    <%= button_link_to Spree.t(:new_gift_card), new_object_url, :icon => 'icon-plus' %>
+  </li>
 <% end %>
 
 <% content_for :table_filter_title do %>
   <h2>Testing</h2>
 <% end %>
 
-<table class="index table table-hover">
+<table class="index">
   <thead>
     <tr>
       <th><%= Spree.t(:code) %></th>
