+++ app/views/spree/admin/gift_cards/new.html.erb
@@ -3,7 +3,9 @@
 <% end %>
 
 <% content_for :page_actions do %>
-  <%= button_link_to Spree.t(:back_to_gift_cards_list), spree.admin_gift_cards_path, :icon => 'arrow-left', class: 'btn btn-primary' %>
+  <li>
+    <%= button_link_to Spree.t(:back_to_gift_cards_list), spree.admin_gift_cards_path, :icon => 'icon-arrow-left' %>
+  </li>
 <% end %>
 
 <%= render :partial => 'spree/shared/error_messages', :locals => { :target => @gift_card } %>
