+++ config/locales/en.yml
@@ -18,5 +18,3 @@ en:
     successfully_created_gift_card: 'You have successfully created the gift card.'
     transactions: "Transactions"
     value: 'Value'
-    gift_card_email:
-        subject: 'Congratulation! Gift Card Received!'
