+++ spree_gift_card.gemspec
@@ -1,28 +1,26 @@
 Gem::Specification.new do |s|
   s.platform    = Gem::Platform::RUBY
   s.name        = 'spree_gift_card'
-  s.version     = '3.0.0.beta'
+  s.version     = '1.0.0.beta'
   s.summary     = 'Spree Gift Card'
   s.description = 'Spree Gift Card Extension'
 
   s.authors     = ['Jeff Dutil']
   s.email       = ['jdutil@burlingtonwebapps.com']
-  s.homepage    = 'https://github.com/vinsol/spree_gift_card'
-
-  spree_version = '~> 3.0.4'
+  s.homepage    = 'http://github.com/jdutil/spree_gift_card'
 
   s.files       = `git ls-files`.split("\n")
   s.test_files  = `git ls-files -- {test,spec,features}/*`.split("\n")
 
   s.require_path = 'lib'
-  s.required_ruby_version = '>= 2.2.0'
+  s.required_ruby_version = '>= 1.9.3'
   s.requirements << 'none'
 
   s.add_dependency 'durable_decorator', '~> 0.2.0'
-  s.add_dependency 'spree_api',         spree_version
-  s.add_dependency 'spree_backend',     spree_version
-  s.add_dependency 'spree_core',        spree_version
-  s.add_dependency 'spree_frontend',    spree_version
+  s.add_dependency 'spree_api',         '~> 2.2.0'
+  s.add_dependency 'spree_backend',     '~> 2.2.0'
+  s.add_dependency 'spree_core',        '~> 2.2.0'
+  s.add_dependency 'spree_frontend',    '~> 2.2.0'
 
   s.add_development_dependency 'capybara', '~> 2.0'
   s.add_development_dependency 'coffee-rails'
